backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.3.2
numpy==2.3.3
plotly==6.3.0
openai==1.107.2
python-multipart==0.0.6
python-dotenv==1.1.1
scikit-learn==1.7.2
pydantic==2.11.9
